move FormatType maps into shared FormatTypeConsts class
