+++ src/tests/Feature/Admin/FormatPartialUpdateTest.php
@@ -28,7 +28,7 @@ class FormatPartialUpdateTest extends TestCase
             if ($fieldName == 'type') {
                 if (!array_key_exists('type', $payload)) {
                     if (!empty($format->format_type_enum)) {
-                        $payload['type'] = FormatUpdateTest::COMDEF_TYPE_TO_TYPE_MAP[$format->format_type_enum];
+                        $payload['type'] = FormatTypeConsts::COMDEF_TYPE_TO_TYPE_MAP[$format->format_type_enum];
                     }
                 }
             }
@@ -113,22 +113,22 @@ class FormatPartialUpdateTest extends TestCase
             $this->assertNull($format->format_type_enum);
         }
 
-        $data = ['type' => FormatUpdateTest::TYPE_MEETING_FORMAT];
+        $data = ['type' => FormatTypeConsts::TYPE_MEETING_FORMAT];
         $this->withHeader('Authorization', "Bearer $token")
             ->patch("/api/v1/formats/{$formats[0]->shared_id_bigint}", $data)
             ->assertStatus(204);
         foreach ($formats as $format) {
             $format = Format::query()->where('shared_id_bigint', $format->shared_id_bigint)->where('lang_enum', $format->lang_enum)->first();
-            $this->assertEquals(FormatUpdateTest::TYPE_TO_COMDEF_TYPE_MAP[$data['type']], $format->format_type_enum);
+            $this->assertEquals(FormatTypeConsts::TYPE_TO_COMDEF_TYPE_MAP[$data['type']], $format->format_type_enum);
         }
 
-        $data = ['type' => FormatUpdateTest::TYPE_MEETING_FORMAT];
+        $data = ['type' => FormatTypeConsts::TYPE_MEETING_FORMAT];
         $this->withHeader('Authorization', "Bearer $token")
             ->patch("/api/v1/formats/{$formats[0]->shared_id_bigint}", $data)
             ->assertStatus(204);
         foreach ($formats as $format) {
             $format = Format::query()->where('shared_id_bigint', $format->shared_id_bigint)->where('lang_enum', $format->lang_enum)->first();
-            $this->assertEquals(FormatUpdateTest::TYPE_TO_COMDEF_TYPE_MAP[$data['type']], $format->format_type_enum);
+            $this->assertEquals(FormatTypeConsts::TYPE_TO_COMDEF_TYPE_MAP[$data['type']], $format->format_type_enum);
         }
     }
 
@@ -140,7 +140,7 @@ class FormatPartialUpdateTest extends TestCase
         $data = $this->toPayload($formats, 'translations');
 
         $data['worldId'] = 'modified';
-        $data['type'] = FormatUpdateTest::TYPE_MEETING_FORMAT;
+        $data['type'] = FormatTypeConsts::TYPE_MEETING_FORMAT;
         foreach ($data['translations'] as $key => $translation) {
             $translation['key'] .= 'updated';
             $translation['name'] .= 'updated';
@@ -156,7 +156,7 @@ class FormatPartialUpdateTest extends TestCase
             $translation = collect($data['translations'])->firstWhere('language', $format->lang_enum);
             $format = Format::query()->where('shared_id_bigint', $format->shared_id_bigint)->where('lang_enum', $format->lang_enum)->first();
             $this->assertEquals($data['worldId'], $format->worldid_mixed);
-            $this->assertEquals(FormatUpdateTest::TYPE_TO_COMDEF_TYPE_MAP[$data['type']], $format->format_type_enum);
+            $this->assertEquals(FormatTypeConsts::TYPE_TO_COMDEF_TYPE_MAP[$data['type']], $format->format_type_enum);
             $this->assertEquals($translation['key'], $format->key_string);
             $this->assertEquals($translation['name'], $format->name_string);
             $this->assertEquals($translation['description'], $format->description_string);
@@ -333,7 +333,7 @@ class FormatPartialUpdateTest extends TestCase
             ->assertStatus(422);
 
         // it can a valid value
-        foreach (array_keys(FormatUpdateTest::TYPE_TO_COMDEF_TYPE_MAP) as $validType) {
+        foreach (array_keys(FormatTypeConsts::TYPE_TO_COMDEF_TYPE_MAP) as $validType) {
             $data['type'] = $validType;
             $this->withHeader('Authorization', "Bearer $token")
                 ->patch("/api/v1/formats/{$formats[0]->shared_id_bigint}", $data)
