+++ src/tests/Feature/Admin/FormatShowTest.php
@@ -63,7 +63,7 @@ class FormatShowTest extends TestCase
             ->json();
 
         $this->assertIsString($data['type']);
-        $this->assertEquals(FormatUpdateTest::COMDEF_TYPE_TO_TYPE_MAP[$format->format_type_enum], $data['type']);
+        $this->assertEquals(FormatTypeConsts::COMDEF_TYPE_TO_TYPE_MAP[$format->format_type_enum], $data['type']);
     }
 
     public function testShowFormatTypeNull()
